requirements.txt
langchain-core
typing-extensions
langgraph
langchain
langchain-openai
langchain-deepseek
pydantic>=2.6.0
pydantic-settings>=2.3.0
openai>=1.35.0
python-dotenv>=1.0.0
PyQt5>=5.15.0
